Add linear variant of MaskedAttention.
This variant's memory usage scales with N instead of N^2, where N is
the sequence length. However, due to the lack of vectorization, it is
much slower to compute. It's possible to vectorize the computation
but this would probably require a further N*D in memory where D is the
embedding dimension.