Return logits from model instead of probabilities.
Also we now return output of shape (N, out_dim*C, H, W) instead of
(N, out_dim, C, H, W) and let the user do any further reshaping. This
may be a bad idea and we might revert this.

Returning logits simplifies the API as the user no longer needs to pass
in a probs_fn. It also decouples the model from the probs generation.

We may also want to decouple sampling from the model entirely as this
would further simplify the API. However, there may be some issues with
this since some models require custom code to efficiently sample (e.g.
NADE, Linear Transformer).

FWIW, it's unclear whether this change is better than the current
approach and we may revert it.